+++ clickhouse_connect/driver/httputil.py
@@ -108,15 +108,14 @@ class ResponseSource:
         self.response = response
         compression = response.headers.get('content-encoding')
         if compression == 'zstd':
-            zstd_decom = zstandard.ZstdDecompressor()
-            reader = zstd_decom.stream_reader(self, read_across_frames=False)
+            zstd_decom = zstandard.ZstdDecompressor().decompressobj()
 
             def decompress():
                 while True:
-                    chunk = reader.read()
+                    chunk = response.read(chunk_size)
                     if not chunk:
                         break
-                    yield chunk
+                    yield zstd_decom.decompress(chunk)
 
             self.gen = decompress()
         elif compression == 'lz4':
@@ -137,11 +136,8 @@ class ResponseSource:
         else:
             self.gen = response.stream(decode_content=True)
 
-    def read(self, amt: int) -> bytes:
-        return self.response.read(amt)
-
     def close(self, ex: Exception = None):
         if ex:
-            logger.warning('Closed HTTP response due to unexpected exception', ex)
+            logger.warning('Closed HTTP response due to unexpected exception %s', str(ex))
         self.response.drain_conn()
         self.response.close()